% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{combine_databases}
\alias{combine_databases}
\title{Create a Geneset Database}
\usage{
combine_databases(databases = "", geneset_names = "")
}
\arguments{
\item{databases}{a list of geneset database files that contain the gene sets}

\item{geneset_names}{a list of lists. The first list corresponds to the first db_file and so on and says what genesets to include in new database. If equal to "all" all genesets in that database will be included}
}
\value{
a dataframe that can be saved to the package or used directly by the GSEAplots function
}
\description{
This function takes a list of geneset databases. It returns a list that can be saved as a new database.
}
