% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.GSEA.022319.R
\name{GSEAplots}
\alias{GSEAplots}
\title{Execute GSEA Analysis}
\usage{
GSEAplots(wd_new = wd_new, datasets.folder = "",
  genesets.folder = "", code.folder = "", input.ds.name = "",
  input.cls.name = "", gene.set.input = "", doc.string = "",
  nperm = 1000, nom.p.val.threshold = -1, fdr.q.val.threshold = 0.25,
  bar_percent = 0.1, gap_percent = 0.1, under_percent = 0.1,
  upper_percent = 0.1, color_line = "black", color_tick = "black")
}
\arguments{
\item{wd_new}{The directory the summary files and pdf of plots will output to}

\item{datasets.folder}{The path to the folder containing expression gct file and phenotype cls file}

\item{genesets.folder}{The path to the folder containing the geneset}

\item{code.folder}{The path to the folder containing the plotting function and GSEA source code}

\item{input.ds.name}{Name of the gct expression file}

\item{input.cls.name}{Name of the cls phenotype file}

\item{gene.set.input}{Name of the geneset file}

\item{doc.string}{Name of the output folder for analysis and for naming output files}

\item{nperm}{number of permutations}

\item{bar_percent}{proportional height of tick mark to window}

\item{gap_percent}{proportional height between minimum ES and tick mark to the window}

\item{under_percent}{proportional height of white space under tick marks to the window size}

\item{upper_percent}{proportional height of white space over enrichment graph to the window size}

\item{color_line}{color of enrichment score line in plot pdf}

\item{color_tick}{color to tick marks on plot}
}
\value{
list pp which includes lists: plots, gene.set.reference.matrix, gene.set.leading, report1, report2, ES
}
\description{
Analyzes relation of differential expression to gene sets
}
\references{
Subramanian, Tamayo, et al. (2005), PNAS 102, 15545-15550, http://www.broad.mit.edu/gsea/
}
